fix: move locales to a separate file to avoid import issues (probably a circular dependency chain?)
