+++ index.ts
@@ -7,6 +7,4 @@
 export { App } from './src/components/app/app';
 export { Login } from './src/components/login/Login';
 export { Error } from './src/components/error/Error';
-export { default as defaultLocale } from './src/resources/scripts/i18n/defaultLocale';
-export { default as informalLocale } from './src/resources/scripts/i18n/informalLocale';
 export * from './src/globalState';
