+++ locales.ts
@@ -1,0 +1,2 @@
+export { default as de } from './src/resources/scripts/i18n/de';
+export { default as deInformal } from './src/resources/scripts/i18n/deInformal';
