+++ src/resources/scripts/i18n/defaultLocale.ts
@@ -1,51 +1,3 @@
-import app from './de/app';
-import absence from './de/absence';
-import attachments from './de/attachments';
-import chatFlyout from './de/chatFlyout';
-import consultant from './de/consultant';
-import deleteAccount from './de/deleteAccount';
-import enquiry from './de/enquiry';
-import furtherSteps from './de/furtherSteps';
-import groupChat from './de/groupChat';
-import login from './de/login';
-import message from './de/message';
-import monitoring from './de/monitoring';
-import navigation from './de/navigation';
-import profile from './de/profile';
-import registration from './de/registration';
-import resort from './de/resort';
-import session from './de/session';
-import sessionList from './de/sessionList';
-import statusOverlay from './de/statusOverlay';
-import typingIndicator from './de/typingIndicator';
-import user from './de/user';
-import userProfile from './de/userProfile';
-import videoCall from './de/videoCall';
+import de from './de';
 
-const defaultLocale = {
-	app,
-	absence,
-	attachments,
-	chatFlyout,
-	consultant,
-	deleteAccount,
-	enquiry,
-	furtherSteps,
-	groupChat,
-	login,
-	message,
-	monitoring,
-	navigation,
-	profile,
-	registration,
-	resort,
-	session,
-	sessionList,
-	statusOverlay,
-	typingIndicator,
-	user,
-	userProfile,
-	videoCall
-};
-
-export default defaultLocale;
+export default de;
